backend/pyproject.toml
[build-system]
requires = ["poetry>=1.0.5"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Łukasz Skarżyński <skarzynski_lukasz@protonmail.com>"]

[tool.poetry.dependencies]
bcrypt = "^3.1"
celery = "^4.3"
django-environ = "^0.4.5"
eventlet = "^0.25.1"
flask = "^1.1"
flask-jwt-extended = "^3.23"
flask-marshmallow = "^0.10.1"
flask-migrate = "^2.5"
flask-socketio = "^4.2"
flask-sqlalchemy = "^2.4"
flower = "^0.9.3"
marshmallow = "^3.2"
marshmallow-sqlalchemy = "^0.19.0"
passlib = "^1.7"
psycopg2-binary = "^2.8"
python = "^3.7"
redis = "^3.3"
sqlalchemy = "^1.3"
sqlalchemy-utils = "^0.34.2"

[tool.poetry.dev-dependencies]
dotenv-linter = "^0.1.4"
factory_boy = "^2.12"
flask-shell-ipython = "^0.4.1"
ipython = "^7.8"
isort = "^4.3"
pytest = "^5.1"
pytest-factoryboy = "^2.0"
pytest-isort = "^0.3.1"
wemake-python-styleguide = "^0.14.0"
yamllint = "^1.17"
flake8-pytest-style = "^1.0.0"
